+++ releases.md
@@ -7,6 +7,7 @@
 - removed `linkedList` module from the stdlib
 - renamed `addFirst` to `::` and `addLast` to `+=` in the core module
 - added `addFirst` and `addLast` functions to the list module
+- added `Pattern` class to the `prim` namespace
 
 ## Bug Fixes